+++ robot-core/src/main/java/org/obolibrary/robot/ExportOperation.java
@@ -1,5 +1,6 @@
 package org.obolibrary.robot;
 
+import com.google.common.collect.Lists;
 import java.io.*;
 import java.util.*;
 import java.util.regex.Matcher;
@@ -19,6 +20,12 @@ public class ExportOperation {
 
   /** Namespace for error messages. */
   private static final String NS = "export#";
+
+  private static final String entitySelectError =
+      NS + "ENTITY SELECT ERROR '%s' is not a valid entity selection";
+
+  private static final String entityFormatError =
+      NS + "ENTITY FORMAT ERROR '%s' is not a valid entity rendering format";
 
   private static final String invalidColumnError =
       NS + "INVALID COLUMN ERROR unable to find property for column header '%s'";
@@ -55,13 +62,18 @@ public class ExportOperation {
   public static Map<String, String> getDefaultOptions() {
     Map<String, String> options = new HashMap<>();
     options.put("include", "classes individuals");
-    options.put("exclude-anonymous", "false");
     options.put("format", "tsv");
     options.put("sort", null);
     options.put("split", "|");
     options.put("entity-format", "NAME");
+    options.put("entity-select", "ANY");
     return options;
   }
+
+  private static final List<String> entityFormatTags =
+      Lists.newArrayList("id", "iri", "label", "name");
+  private static final List<String> entitySelectTags =
+      Lists.newArrayList("any", "named", "anon", "anonymous");
 
   /**
    * Given an ontology, an ioHelper, a list of columns, an output export file, and a map of options,
@@ -80,8 +92,6 @@ public class ExportOperation {
       List<String> columnNames,
       Map<String, String> options)
       throws Exception {
-    boolean excludeAnonymous = OptionsHelper.optionIsTrue(options, "exclude-anonymous");
-
     // Get column or columns to sort on
     // If not provided, use the first column
     String sortColumn = OptionsHelper.getOption(options, "sort", columnNames.get(0));
@@ -105,6 +115,13 @@ public class ExportOperation {
 
     // Default tag for entity rendering when not specified
     String entityFormat = OptionsHelper.getOption(options, "entity-format", "NAME");
+    if (!entityFormatTags.contains(entityFormat.toLowerCase())) {
+      throw new Exception(String.format(entityFormatError, entityFormat));
+    }
+    String entitySelect = OptionsHelper.getOption(options, "entity-select", "ANY");
+    if (!entitySelectTags.contains(entitySelect.toLowerCase())) {
+      throw new Exception(String.format(entitySelectError, entitySelect));
+    }
 
     // Create some providers for rendering entities
     ShortFormProvider oboProvider = new CURIEShortFormProvider(ioHelper.getPrefixes());
@@ -127,13 +144,28 @@ public class ExportOperation {
     // Create the Column objects and add to table
     List<String> sorts = Arrays.asList(sortColumn.trim().split("\\|"));
     for (String c : columnNames) {
-      String tag = entityFormat;
+      String currentEntityFormat = entityFormat;
+      String currentEntitySelect = entitySelect;
       String colName = c;
       // Determine if this has a tag for rendering
-      Matcher m = Pattern.compile("(.+) \\[(ID|id|IRI|iri|LABEL|label|NAME|name)]").matcher(c);
+      Matcher m =
+          Pattern.compile(
+                  "(.+) \\[(id|iri|label|named|name|anon|anonymous|any) ?.*]",
+                  Pattern.CASE_INSENSITIVE)
+              .matcher(c);
       if (m.find()) {
         colName = m.group(1);
-        tag = m.group(2);
+        String tag = m.group(2);
+        // Process one or more tags
+        for (String subTag : tag.split(" ")) {
+          if (entityFormatTags.contains(subTag.toLowerCase())) {
+            currentEntityFormat = subTag;
+          } else if (entitySelectTags.contains(subTag.toLowerCase())) {
+            currentEntitySelect = subTag;
+          } else {
+            throw new Exception(String.format(unknownTagError, c, subTag));
+          }
+        }
       }
 
       // Add some other defaults to the label map
@@ -147,21 +179,34 @@ public class ExportOperation {
 
       // Handle the default column rendering
       if (c.equalsIgnoreCase("ID") || c.equalsIgnoreCase("CURIE")) {
-        tag = "ID";
+        currentEntityFormat = "ID";
+        currentEntitySelect = "NAMED";
       } else if (c.equalsIgnoreCase("IRI")) {
-        tag = "IRI";
+        currentEntityFormat = "IRI";
+        currentEntitySelect = "NAMED";
       } else if (c.equalsIgnoreCase("LABEL")) {
-        tag = "LABEL";
+        currentEntityFormat = "LABEL";
+        currentEntitySelect = "NAMED";
       }
 
       // Maybe get a property
       OWLAnnotationProperty ap = checker.getOWLAnnotationProperty(colName, false);
+
+      // Handle some defaults
+      IRI colIRI = ioHelper.createIRI(colName);
+      if (colIRI != null
+          && colIRI.toString().equals(dataFactory.getRDFSLabel().getIRI().toString())) {
+        currentEntityFormat = "LABEL";
+        currentEntitySelect = "NAMED";
+        ap = dataFactory.getRDFSLabel();
+      }
+
       OWLDataProperty dp = checker.getOWLDataProperty(colName);
       OWLObjectProperty op = checker.getOWLObjectProperty(colName);
 
       // Maybe get a short form provider
       ShortFormProvider provider;
-      switch (tag.toUpperCase()) {
+      switch (currentEntityFormat.toUpperCase()) {
         case "ID":
         case "CURIE":
           provider = oboProvider;
@@ -176,18 +221,18 @@ public class ExportOperation {
           provider = labelProvider;
           break;
         default:
-          throw new Exception(String.format(unknownTagError, c, tag));
+          throw new Exception(String.format(unknownTagError, c, currentEntityFormat));
       }
 
       Column column;
       if (ap != null) {
         column = new Column(colName, c, ap, provider);
       } else if (dp != null) {
-        column = new Column(colName, c, dp, provider);
+        column = new Column(colName, c, dp, provider, currentEntitySelect);
       } else if (op != null) {
-        column = new Column(colName, c, op, provider);
+        column = new Column(colName, c, op, provider, currentEntitySelect);
       } else {
-        column = new Column(colName, c, iri, provider);
+        column = new Column(colName, c, iri, provider, currentEntitySelect);
       }
 
       // Maybe set sort info
@@ -212,7 +257,7 @@ public class ExportOperation {
 
     // Get the cell values based on columns
     for (OWLEntity entity : entities) {
-      table.addRow(getRow(ontology, table, entity, excludeAnonymous));
+      table.addRow(getRow(ontology, table, entity));
     }
 
     // Sort the rows by sort column or columns
@@ -269,21 +314,23 @@ public class ExportOperation {
    * @param rt RendererType to use to render Manchester
    * @param provider ShortFormProvider to resolve entities
    * @param classes Set of class expressions to convert to string
-   * @param excludeAnonymous if true, exclude anonymous class expressions
+   * @param includeNamed
+   * @param includeAnonymous
    * @return String of class expressions or null
    */
   private static List<String> classExpressionsToString(
       RendererType rt,
       ShortFormProvider provider,
       Collection<OWLClassExpression> classes,
-      boolean excludeAnonymous) {
+      boolean includeNamed,
+      boolean includeAnonymous) {
     List<String> strings = new ArrayList<>();
     for (OWLClassExpression expr : classes) {
-      if (expr.isAnonymous() && !excludeAnonymous) {
+      if (expr.isAnonymous() && includeAnonymous) {
         // Get a Manchester string using labels
         String manString = renderManchester(rt, provider, expr);
         strings.add(manString);
-      } else if (!expr.isAnonymous()) {
+      } else if (!expr.isAnonymous() && includeNamed) {
         OWLClass sc = expr.asOWLClass();
         strings.add(renderManchester(rt, provider, sc));
       }
@@ -299,7 +346,8 @@ public class ExportOperation {
    * @param displayRendererType RendererType for display value
    * @param sortRendererType RendererType for sort value
    * @param provider ShortFormProvider to resolve entities
-   * @param excludeAnonymous if true, exclude anonymous expressions
+   * @param includeNamed
+   * @param includeAnonymous
    * @return Cell for this Column containing class expressions
    */
   private static Cell getClassCell(
@@ -308,12 +356,16 @@ public class ExportOperation {
       RendererType displayRendererType,
       RendererType sortRendererType,
       ShortFormProvider provider,
-      boolean excludeAnonymous) {
+      boolean includeNamed,
+      boolean includeAnonymous) {
     List<String> displays =
-        classExpressionsToString(displayRendererType, provider, exprs, excludeAnonymous);
+        classExpressionsToString(
+            displayRendererType, provider, exprs, includeNamed, includeAnonymous);
     List<String> sorts;
     if (sortRendererType != null) {
-      sorts = classExpressionsToString(sortRendererType, provider, exprs, excludeAnonymous);
+      sorts =
+          classExpressionsToString(
+              sortRendererType, provider, exprs, includeNamed, includeAnonymous);
     } else {
       sorts = displays;
     }
@@ -393,6 +445,28 @@ public class ExportOperation {
   }
 
   /**
+   * Return a cell containing entity type, rendered based on the provider for the Column.
+   *
+   * @param type EntityType of the target entity
+   * @param column Column for this cell
+   * @return Cell for this Column containing entity type as string rendering
+   */
+  private static Cell getEntityTypeCell(EntityType type, Column column) {
+    ShortFormProvider provider = column.getShortFormProvider();
+    String cellValue;
+    if (provider instanceof CURIEShortFormProvider) {
+      CURIEShortFormProvider sfp = (CURIEShortFormProvider) provider;
+      cellValue = sfp.getShortForm(type.getIRI());
+    } else if (provider instanceof AnnotationValueShortFormProvider) {
+      cellValue = type.getPrintName();
+    } else {
+      // IRI provider
+      cellValue = type.getIRI().toString();
+    }
+    return new Cell(column, cellValue);
+  }
+
+  /**
    * Return a Cell containing property expressions.
    *
    * @param exprs Collection of OWLPropertyExpressions to render in property cell
@@ -400,7 +474,8 @@ public class ExportOperation {
    * @param displayRendererType RendererType for display value
    * @param sortRendererType RendererType for sort value
    * @param provider ShortFormProvider to resolve entities
-   * @param excludeAnonymous if true, exclude anonymous expressions
+   * @param includeNamed
+   * @param includeAnonymous
    * @return Cell for this Column containing property expressions
    */
   private static Cell getPropertyCell(
@@ -409,12 +484,16 @@ public class ExportOperation {
       RendererType displayRendererType,
       RendererType sortRendererType,
       ShortFormProvider provider,
-      boolean excludeAnonymous) {
+      boolean includeNamed,
+      boolean includeAnonymous) {
     List<String> displays =
-        propertyExpressionsToString(displayRendererType, provider, exprs, excludeAnonymous);
+        propertyExpressionsToString(
+            displayRendererType, provider, exprs, includeNamed, includeAnonymous);
     List<String> sorts;
     if (sortRendererType != null) {
-      sorts = propertyExpressionsToString(sortRendererType, provider, exprs, excludeAnonymous);
+      sorts =
+          propertyExpressionsToString(
+              sortRendererType, provider, exprs, includeNamed, includeAnonymous);
     } else {
       sorts = displays;
     }
@@ -442,7 +521,17 @@ public class ExportOperation {
     for (OWLAnnotationAssertionAxiom a :
         EntitySearcher.getAnnotationAssertionAxioms(entity, ontology)) {
       if (a.getProperty().getIRI() == ap.getIRI()) {
-        values.add(renderManchester(rt, provider, a.getValue()));
+        if (a.getValue().isIRI()) {
+          IRI iri = a.getValue().asIRI().orNull();
+          if (iri != null) {
+            Set<OWLEntity> entities = ontology.getEntitiesInSignature(iri);
+            for (OWLEntity e : entities) {
+              values.add(renderManchester(rt, provider, e));
+            }
+          }
+        } else {
+          values.add(renderManchester(rt, provider, a.getValue()));
+        }
       }
     }
     return values;
@@ -458,7 +547,8 @@ public class ExportOperation {
    * @param provider ShortFormProvider to resolve entities
    * @param entity OWLEntity to get relations of
    * @param dp OWLDataProperty to get the value(s) of
-   * @param excludeAnonymous if true, do not include anonymous class expressions
+   * @param includeNamed
+   * @param includeAnonymous
    * @return String of values or null
    */
   private static List<String> getPropertyValues(
@@ -467,7 +557,8 @@ public class ExportOperation {
       ShortFormProvider provider,
       OWLEntity entity,
       OWLDataProperty dp,
-      boolean excludeAnonymous) {
+      boolean includeNamed,
+      boolean includeAnonymous) {
     if (entity.isOWLNamedIndividual()) {
       OWLNamedIndividual i = entity.asOWLNamedIndividual();
       Collection<OWLLiteral> propVals = EntitySearcher.getDataPropertyValues(i, dp, ontology);
@@ -482,7 +573,8 @@ public class ExportOperation {
         }
         // break down into conjuncts
         vals.addAll(
-            getRestrictionFillers(expr.asConjunctSet(), dp, rt, provider, excludeAnonymous));
+            getRestrictionFillers(
+                expr.asConjunctSet(), dp, rt, provider, includeNamed, includeAnonymous));
       }
       // Find equivalent class expressions that use this property
       for (OWLClassExpression expr :
@@ -492,7 +584,8 @@ public class ExportOperation {
         }
         // break down into conjuncts
         vals.addAll(
-            getRestrictionFillers(expr.asConjunctSet(), dp, rt, provider, excludeAnonymous));
+            getRestrictionFillers(
+                expr.asConjunctSet(), dp, rt, provider, includeNamed, includeAnonymous));
       }
       return vals;
     } else {
@@ -510,7 +603,8 @@ public class ExportOperation {
    * @param provider ShortFormProvider to resolve entities
    * @param entity OWLEntity to get annotations on
    * @param op OWLObjectProperty to get the value(s) of
-   * @param excludeAnonymous if true, do not include anonymous class expressions
+   * @param includeNamed
+   * @param includeAnonymous
    * @return String of values or null
    */
   private static List<String> getPropertyValues(
@@ -519,7 +613,8 @@ public class ExportOperation {
       ShortFormProvider provider,
       OWLEntity entity,
       OWLObjectProperty op,
-      boolean excludeAnonymous) {
+      boolean includeNamed,
+      boolean includeAnonymous) {
     if (entity.isOWLNamedIndividual()) {
       OWLNamedIndividual i = entity.asOWLNamedIndividual();
       Collection<OWLIndividual> propVals = EntitySearcher.getObjectPropertyValues(i, op, ontology);
@@ -538,7 +633,8 @@ public class ExportOperation {
         }
         // break down into conjuncts
         exprs.addAll(
-            getRestrictionFillers(expr.asConjunctSet(), op, rt, provider, excludeAnonymous));
+            getRestrictionFillers(
+                expr.asConjunctSet(), op, rt, provider, includeNamed, includeAnonymous));
       }
       // Find equivalent class expressions that use this property
       for (OWLClassExpression expr :
@@ -548,7 +644,8 @@ public class ExportOperation {
         }
         // break down into conjuncts
         exprs.addAll(
-            getRestrictionFillers(expr.asConjunctSet(), op, rt, provider, excludeAnonymous));
+            getRestrictionFillers(
+                expr.asConjunctSet(), op, rt, provider, includeNamed, includeAnonymous));
       }
       return exprs;
     } else {
@@ -566,7 +663,8 @@ public class ExportOperation {
    * @param dp OWLDataProperty to look for
    * @param rt RendererType to use to render Manchester
    * @param provider ShortFormProvider to resolve entities
-   * @param excludeAnonymous if true, exclude anonymous data ranges
+   * @param includeNamed
+   * @param includeAnonymous
    * @return set of fillers that are 'values' of the data property
    */
   private static Set<String> getRestrictionFillers(
@@ -574,7 +672,8 @@ public class ExportOperation {
       OWLDataProperty dp,
       RendererType rt,
       ShortFormProvider provider,
-      boolean excludeAnonymous) {
+      boolean includeNamed,
+      boolean includeAnonymous) {
     Set<String> fillers = new HashSet<>();
     for (OWLClassExpression ce : exprs) {
       // Determine the type of restriction
@@ -587,13 +686,12 @@ public class ExportOperation {
           OWLDataAllValuesFrom avf = (OWLDataAllValuesFrom) ce;
           f = avf.getFiller();
           pe = avf.getProperty();
-          if (excludeAnonymous && f.isAnonymous()) {
-            continue;
-          }
           if (!pe.isAnonymous()) {
             OWLDataProperty prop = pe.asOWLDataProperty();
             if (prop.getIRI() == dp.getIRI()) {
-              fillers.add(renderRestrictionString(rt, provider, f, null));
+              if ((!f.isAnonymous() && includeNamed) || (f.isAnonymous() && includeAnonymous)) {
+                fillers.add(renderRestrictionString(rt, provider, f, null));
+              }
             }
           }
           break;
@@ -601,13 +699,12 @@ public class ExportOperation {
           OWLDataSomeValuesFrom svf = (OWLDataSomeValuesFrom) ce;
           f = svf.getFiller();
           pe = svf.getProperty();
-          if (excludeAnonymous && f.isAnonymous()) {
-            continue;
-          }
           if (!pe.isAnonymous()) {
             OWLDataProperty prop = pe.asOWLDataProperty();
             if (prop.getIRI() == dp.getIRI()) {
-              fillers.add(renderRestrictionString(rt, provider, f, null));
+              if ((!f.isAnonymous() && includeNamed) || (f.isAnonymous() && includeAnonymous)) {
+                fillers.add(renderRestrictionString(rt, provider, f, null));
+              }
             }
           }
           break;
@@ -616,13 +713,12 @@ public class ExportOperation {
           f = ec.getFiller();
           pe = ec.getProperty();
           n = ec.getCardinality();
-          if (excludeAnonymous && f.isAnonymous()) {
-            continue;
-          }
           if (!pe.isAnonymous()) {
             OWLDataProperty prop = pe.asOWLDataProperty();
             if (prop.getIRI() == dp.getIRI()) {
-              fillers.add(renderRestrictionString(rt, provider, f, n));
+              if ((!f.isAnonymous() && includeNamed) || (f.isAnonymous() && includeAnonymous)) {
+                fillers.add(renderRestrictionString(rt, provider, f, n));
+              }
             }
           }
           break;
@@ -631,13 +727,12 @@ public class ExportOperation {
           f = minc.getFiller();
           pe = minc.getProperty();
           n = minc.getCardinality();
-          if (excludeAnonymous && f.isAnonymous()) {
-            continue;
-          }
           if (!pe.isAnonymous()) {
             OWLDataProperty prop = pe.asOWLDataProperty();
             if (prop.getIRI() == dp.getIRI()) {
-              fillers.add(renderRestrictionString(rt, provider, f, n));
+              if ((!f.isAnonymous() && includeNamed) || (f.isAnonymous() && includeAnonymous)) {
+                fillers.add(renderRestrictionString(rt, provider, f, n));
+              }
             }
           }
           break;
@@ -646,13 +741,12 @@ public class ExportOperation {
           f = maxc.getFiller();
           pe = maxc.getProperty();
           n = maxc.getCardinality();
-          if (excludeAnonymous && f.isAnonymous()) {
-            continue;
-          }
           if (!pe.isAnonymous()) {
             OWLDataProperty prop = pe.asOWLDataProperty();
             if (prop.getIRI() == dp.getIRI()) {
-              fillers.add(renderRestrictionString(rt, provider, f, n));
+              if ((!f.isAnonymous() && includeNamed) || (f.isAnonymous() && includeAnonymous)) {
+                fillers.add(renderRestrictionString(rt, provider, f, n));
+              }
             }
           }
           break;
@@ -671,7 +765,8 @@ public class ExportOperation {
    * @param op OWLObjectProperty to look for
    * @param rt RendererType to use to render Manchester
    * @param provider ShortFormProvider to resolve entities
-   * @param excludeAnonymous if true, exclude anonymous class expressions
+   * @param includeNamed
+   * @param includeAnonymous
    * @return set of fillers that are 'values' of the object property
    */
   private static Set<String> getRestrictionFillers(
@@ -679,7 +774,8 @@ public class ExportOperation {
       OWLObjectProperty op,
       RendererType rt,
       ShortFormProvider provider,
-      boolean excludeAnonymous) {
+      boolean includeNamed,
+      boolean includeAnonymous) {
     Set<String> fillers = new HashSet<>();
     for (OWLClassExpression ce : exprs) {
       // Determine the type of restriction
@@ -693,13 +789,12 @@ public class ExportOperation {
           OWLObjectAllValuesFrom avf = (OWLObjectAllValuesFrom) ce;
           pe = avf.getProperty();
           f = avf.getFiller();
-          if (excludeAnonymous && f.isAnonymous()) {
-            continue;
-          }
           if (!pe.isAnonymous()) {
             OWLObjectProperty prop = pe.asOWLObjectProperty();
             if (prop.getIRI() == op.getIRI()) {
-              fillers.add(renderRestrictionString(rt, provider, f, null));
+              if ((f.isAnonymous() && includeAnonymous) || (!f.isAnonymous() && includeNamed)) {
+                fillers.add(renderRestrictionString(rt, provider, f, null));
+              }
             }
           }
           break;
@@ -708,13 +803,12 @@ public class ExportOperation {
           OWLObjectSomeValuesFrom svf = (OWLObjectSomeValuesFrom) ce;
           pe = svf.getProperty();
           f = svf.getFiller();
-          if (excludeAnonymous && f.isAnonymous()) {
-            continue;
-          }
           if (!pe.isAnonymous()) {
             OWLObjectProperty prop = pe.asOWLObjectProperty();
             if (prop.getIRI() == op.getIRI()) {
-              fillers.add(renderRestrictionString(rt, provider, f, null));
+              if ((f.isAnonymous() && includeAnonymous) || (!f.isAnonymous() && includeNamed)) {
+                fillers.add(renderRestrictionString(rt, provider, f, null));
+              }
             }
           }
           break;
@@ -724,13 +818,12 @@ public class ExportOperation {
           pe = ec.getProperty();
           f = ec.getFiller();
           n = ec.getCardinality();
-          if (excludeAnonymous && f.isAnonymous()) {
-            continue;
-          }
           if (!pe.isAnonymous()) {
             OWLObjectProperty prop = pe.asOWLObjectProperty();
             if (prop.getIRI() == op.getIRI()) {
-              fillers.add(renderRestrictionString(rt, provider, f, n));
+              if ((f.isAnonymous() && includeAnonymous) || (!f.isAnonymous() && includeNamed)) {
+                fillers.add(renderRestrictionString(rt, provider, f, n));
+              }
             }
           }
           break;
@@ -740,13 +833,12 @@ public class ExportOperation {
           pe = minc.getProperty();
           f = minc.getFiller();
           n = minc.getCardinality();
-          if (excludeAnonymous && f.isAnonymous()) {
-            continue;
-          }
           if (!pe.isAnonymous()) {
             OWLObjectProperty prop = pe.asOWLObjectProperty();
             if (prop.getIRI() == op.getIRI()) {
-              fillers.add(renderRestrictionString(rt, provider, f, n));
+              if ((f.isAnonymous() && includeAnonymous) || (!f.isAnonymous() && includeNamed)) {
+                fillers.add(renderRestrictionString(rt, provider, f, n));
+              }
             }
           }
           break;
@@ -756,24 +848,19 @@ public class ExportOperation {
           pe = maxc.getProperty();
           f = maxc.getFiller();
           n = maxc.getCardinality();
-          if (excludeAnonymous && f.isAnonymous()) {
-            continue;
-          }
           if (!pe.isAnonymous()) {
             OWLObjectProperty prop = pe.asOWLObjectProperty();
             if (prop.getIRI() == op.getIRI()) {
-              fillers.add(renderRestrictionString(rt, provider, f, n));
+              if ((f.isAnonymous() && includeAnonymous) || (!f.isAnonymous() && includeNamed)) {
+                fillers.add(renderRestrictionString(rt, provider, f, n));
+              }
             }
           }
           break;
         case OBJECT_ONE_OF:
-          System.out.println("ONE OF " + ce);
-          break;
         case OBJECT_UNION_OF:
-          System.out.println("UNION OF " + ce);
-          break;
         case OBJECT_INTERSECTION_OF:
-          System.out.println("INTERSECTION OF " + ce);
+          // TODO
           break;
       }
     }
@@ -795,13 +882,10 @@ public class ExportOperation {
    * @param ontology OWLOntology to get details from
    * @param table Table to get rendering information and columns
    * @param entity OWLEntity to get details of
-   * @param excludeAnonymous if true, exclude anonymous expressions
    * @return Row object for the OWLEntity
    * @throws Exception on invalid column
    */
-  private static Row getRow(
-      OWLOntology ontology, Table table, OWLEntity entity, boolean excludeAnonymous)
-      throws Exception {
+  private static Row getRow(OWLOntology ontology, Table table, OWLEntity entity) throws Exception {
 
     String format = table.getFormat();
 
@@ -811,7 +895,18 @@ public class ExportOperation {
     Row row = new Row(entity.getIRI());
     for (Column col : table.getColumns()) {
 
+      boolean includeNamed = col.getIncludeNamed();
+      boolean includeAnonymous = col.getIncludeAnonymous();
+
       String colName = col.getName();
+      OWLProperty maybeAnnotation = col.getProperty();
+      if (maybeAnnotation instanceof OWLAnnotationProperty) {
+        OWLAnnotationProperty maybeLabel = (OWLAnnotationProperty) maybeAnnotation;
+        if (maybeLabel.isLabel()) {
+          // Handle like we do default LABEL columns
+          colName = "LABEL";
+        }
+      }
       ShortFormProvider provider = col.getShortFormProvider();
 
       Cell cell;
@@ -852,7 +947,8 @@ public class ExportOperation {
                     displayRendererType,
                     sortRendererType,
                     provider,
-                    excludeAnonymous));
+                    includeNamed,
+                    includeAnonymous));
           }
           continue;
       }
@@ -877,12 +973,24 @@ public class ExportOperation {
           OWLDataProperty dp = (OWLDataProperty) colProperty;
           List<String> display =
               getPropertyValues(
-                  ontology, displayRendererType, provider, entity, dp, excludeAnonymous);
+                  ontology,
+                  displayRendererType,
+                  provider,
+                  entity,
+                  dp,
+                  includeNamed,
+                  includeAnonymous);
           List<String> sort;
           if (sortRendererType != null) {
             sort =
                 getPropertyValues(
-                    ontology, sortRendererType, provider, entity, dp, excludeAnonymous);
+                    ontology,
+                    sortRendererType,
+                    provider,
+                    entity,
+                    dp,
+                    includeNamed,
+                    includeAnonymous);
           } else {
             sort = display;
           }
@@ -894,12 +1002,24 @@ public class ExportOperation {
           OWLObjectProperty op = (OWLObjectProperty) colProperty;
           List<String> display =
               getPropertyValues(
-                  ontology, displayRendererType, provider, entity, op, excludeAnonymous);
+                  ontology,
+                  displayRendererType,
+                  provider,
+                  entity,
+                  op,
+                  includeNamed,
+                  includeAnonymous);
           List<String> sort;
           if (sortRendererType != null) {
             sort =
                 getPropertyValues(
-                    ontology, sortRendererType, provider, entity, op, excludeAnonymous);
+                    ontology,
+                    sortRendererType,
+                    provider,
+                    entity,
+                    op,
+                    includeNamed,
+                    includeAnonymous);
           } else {
             sort = display;
           }
@@ -934,7 +1054,8 @@ public class ExportOperation {
                     displayRendererType,
                     sortRendererType,
                     provider,
-                    excludeAnonymous));
+                    includeNamed,
+                    includeAnonymous));
           }
           break;
         case "http://www.w3.org/2000/01/rdf-schema#subPropertyOf":
@@ -950,7 +1071,8 @@ public class ExportOperation {
                     displayRendererType,
                     sortRendererType,
                     provider,
-                    excludeAnonymous));
+                    includeNamed,
+                    includeAnonymous));
 
           } else if (entity.isOWLDataProperty()) {
             Collection<OWLDataPropertyExpression> supers =
@@ -962,7 +1084,8 @@ public class ExportOperation {
                     displayRendererType,
                     sortRendererType,
                     provider,
-                    excludeAnonymous));
+                    includeNamed,
+                    includeAnonymous));
 
           } else if (entity.isOWLObjectProperty()) {
             Collection<OWLObjectPropertyExpression> supers =
@@ -974,7 +1097,8 @@ public class ExportOperation {
                     displayRendererType,
                     sortRendererType,
                     provider,
-                    excludeAnonymous));
+                    includeNamed,
+                    includeAnonymous));
           }
           break;
         case "http://www.w3.org/2002/07/owl#equivalentClass":
@@ -984,7 +1108,13 @@ public class ExportOperation {
                 EntitySearcher.getEquivalentClasses(entity.asOWLClass(), ontology);
             row.add(
                 getClassCell(
-                    eqs, col, displayRendererType, sortRendererType, provider, excludeAnonymous));
+                    eqs,
+                    col,
+                    displayRendererType,
+                    sortRendererType,
+                    provider,
+                    includeNamed,
+                    includeAnonymous));
           }
           break;
         case "http://www.w3.org/2002/07/owl#equivalentProperty":
@@ -994,21 +1124,39 @@ public class ExportOperation {
                 EntitySearcher.getEquivalentProperties(entity.asOWLAnnotationProperty(), ontology);
             row.add(
                 getPropertyCell(
-                    eqs, col, displayRendererType, sortRendererType, provider, excludeAnonymous));
+                    eqs,
+                    col,
+                    displayRendererType,
+                    sortRendererType,
+                    provider,
+                    includeNamed,
+                    includeAnonymous));
 
           } else if (entity.isOWLDataProperty()) {
             Collection<OWLDataPropertyExpression> eqs =
                 EntitySearcher.getEquivalentProperties(entity.asOWLDataProperty(), ontology);
             row.add(
                 getPropertyCell(
-                    eqs, col, displayRendererType, sortRendererType, provider, excludeAnonymous));
+                    eqs,
+                    col,
+                    displayRendererType,
+                    sortRendererType,
+                    provider,
+                    includeNamed,
+                    includeAnonymous));
 
           } else if (entity.isOWLObjectProperty()) {
             Collection<OWLObjectPropertyExpression> eqs =
                 EntitySearcher.getEquivalentProperties(entity.asOWLObjectProperty(), ontology);
             row.add(
                 getPropertyCell(
-                    eqs, col, displayRendererType, sortRendererType, provider, excludeAnonymous));
+                    eqs,
+                    col,
+                    displayRendererType,
+                    sortRendererType,
+                    provider,
+                    includeNamed,
+                    includeAnonymous));
           }
           break;
         case "http://www.w3.org/2002/07/owl#disjointWith":
@@ -1023,7 +1171,8 @@ public class ExportOperation {
                     displayRendererType,
                     sortRendererType,
                     provider,
-                    excludeAnonymous));
+                    includeNamed,
+                    includeAnonymous));
 
           } else if (entity.isOWLAnnotationProperty()) {
             Collection<OWLAnnotationProperty> disjoints =
@@ -1035,7 +1184,8 @@ public class ExportOperation {
                     displayRendererType,
                     sortRendererType,
                     provider,
-                    excludeAnonymous));
+                    includeNamed,
+                    includeAnonymous));
 
           } else if (entity.isOWLDataProperty()) {
             Collection<OWLDataPropertyExpression> disjoints =
@@ -1047,7 +1197,8 @@ public class ExportOperation {
                     displayRendererType,
                     sortRendererType,
                     provider,
-                    excludeAnonymous));
+                    includeNamed,
+                    includeAnonymous));
 
           } else if (entity.isOWLObjectProperty()) {
             Collection<OWLObjectPropertyExpression> disjoints =
@@ -1059,7 +1210,8 @@ public class ExportOperation {
                     displayRendererType,
                     sortRendererType,
                     provider,
-                    excludeAnonymous));
+                    includeNamed,
+                    includeAnonymous));
           }
           break;
         case "http://www.w3.org/1999/02/22-rdf-syntax-ns#type":
@@ -1067,9 +1219,24 @@ public class ExportOperation {
           if (entity.isOWLNamedIndividual()) {
             Collection<OWLClassExpression> types =
                 EntitySearcher.getTypes(entity.asOWLNamedIndividual(), ontology);
-            row.add(
-                getClassCell(
-                    types, col, displayRendererType, sortRendererType, provider, excludeAnonymous));
+            if (!types.isEmpty()) {
+              row.add(
+                  getClassCell(
+                      types,
+                      col,
+                      displayRendererType,
+                      sortRendererType,
+                      provider,
+                      includeNamed,
+                      includeAnonymous));
+            } else {
+              // No class assertions, just provide the entity type
+              row.add(getEntityTypeCell(entity.getEntityType(), col));
+            }
+
+          } else {
+            // Not an individual, just return the entity type
+            row.add(getEntityTypeCell(entity.getEntityType(), col));
           }
           break;
         default:
@@ -1111,11 +1278,16 @@ public class ExportOperation {
    * @param rt RendererType to use to render Manchester
    * @param provider ShortFormProvider to resolve entities
    * @param props Set of property expressions to convert to string
-   * @param excludeAnonymous if true, exclude anonymous expressions
+   * @param includeNamed
+   * @param includeAnonymous
    * @return String of property expressions or null
    */
   private static List<String> propertyExpressionsToString(
-      RendererType rt, ShortFormProvider provider, Collection<?> props, boolean excludeAnonymous) {
+      RendererType rt,
+      ShortFormProvider provider,
+      Collection<?> props,
+      boolean includeNamed,
+      boolean includeAnonymous) {
     // Try to convert to object property expressions
     Collection<OWLObjectPropertyExpression> opes =
         props
@@ -1137,20 +1309,20 @@ public class ExportOperation {
     // Only one of the above collections will have entries
     // Maybe process object property expressions
     for (OWLObjectPropertyExpression expr : opes) {
-      if (expr.isAnonymous() && !excludeAnonymous) {
+      if (expr.isAnonymous() && includeAnonymous) {
         String manString = renderManchester(rt, provider, expr);
         strings.add(manString);
-      } else if (!expr.isAnonymous()) {
+      } else if (!expr.isAnonymous() && includeNamed) {
         OWLObjectProperty op = expr.asOWLObjectProperty();
         strings.add(renderManchester(rt, provider, op));
       }
     }
     // Maybe process data property expressions
     for (OWLDataPropertyExpression expr : dpes) {
-      if (expr.isAnonymous() && !excludeAnonymous) {
+      if (expr.isAnonymous() && includeAnonymous) {
         String manString = renderManchester(rt, provider, expr);
         strings.add(manString);
-      } else if (!expr.isAnonymous()) {
+      } else if (!expr.isAnonymous() && includeNamed) {
         OWLDataProperty dp = expr.asOWLDataProperty();
         strings.add(renderManchester(rt, provider, dp));
       }